move vue files into vue dir
