+++ kolibri/core/assets/src/vue/content-renderer/index.vue
