+++ kolibri/core/assets/src/vue/core-base.vue
